requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.6.2            # via -r requirements.in, discord.py, twitchio
async-timeout==3.0.1      # via aiohttp, twitchio
attrs==19.3.0             # via aiohttp
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via aiohttp, requests
click==7.1.1              # via pip-tools
discord.py==1.3.2         # via -r requirements.in
google-api-core==1.16.0   # via google-api-python-client
google-api-python-client==1.8.0  # via -r requirements.in
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.12.0       # via google-api-core, google-api-python-client, google-auth-httplib2
googleapis-common-protos==1.51.0  # via google-api-core
httplib2==0.17.0          # via google-api-python-client, google-auth-httplib2, oauth2client
idna==2.9                 # via requests, yarl
multidict==4.7.5          # via aiohttp, yarl
oauth2client==4.1.3       # via -r requirements.in
pip-tools==4.5.1          # via -r requirements.in
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth, oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pytz==2019.3              # via google-api-core
pyyaml==5.3.1             # via -r requirements.in
requests==2.23.0          # via google-api-core
rsa==4.0                  # via google-auth, oauth2client
six==1.14.0               # via google-api-core, google-api-python-client, google-auth, oauth2client, pip-tools, protobuf
twitchio==1.0.0           # via -r requirements.in
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.8           # via requests
websockets==8.1           # via discord.py, twitchio
yarl==1.4.2               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
